feat: refactor authentication tests and add new test_auth module
