+++ accounts/views.py
@@ -1,5 +1,6 @@
 from rest_framework import generics, permissions
 from rest_framework.response import Response
+from drf_spectacular.utils import extend_schema
 from .serializers import UserSerializer
 
 class RegisterView(generics.CreateAPIView):
@@ -7,6 +8,7 @@ class RegisterView(generics.CreateAPIView):
     permission_classes = [permissions.AllowAny]
 
 
+@extend_schema(exclude=True)
 class ProtectedHelloView(generics.GenericAPIView):
     permission_classes = [permissions.IsAuthenticated]
